ref(epidemio): event paciente fallecido
