+++ modules/rup/controllers/elementos-rup.controller.ts
@@ -1,12 +1,6 @@
 import { ElementoRUP, IElementoRUPDoc } from '../schemas/elementoRUP';
 import { Types } from 'mongoose';
 import { AppCache } from '../../../connections';
-import { EventCore } from '@andes/event-bus/';
-import { SeguimientoPaciente } from '../../../modules/seguimiento-paciente/schemas/seguimiento-paciente.schema';
-import { SeguimientoPacienteCtr } from '../../../modules/seguimiento-paciente/seguimiento-paciente.route';
-import { userScheduler } from '../../../config.private';
-
-const dataLog: any = new Object(userScheduler);
 
 export type IElementoRUPDocExt = IElementoRUPDoc & { requeridosMap: { [key: string]: any } };
 
@@ -113,21 +107,3 @@ export async function fulfillPrestacion(prestacion, elementosRUPSet: ElementoRUP
     traverseRegistros(prestacion.ejecucion.registros, elementoPrestacion);
 
 }
-
-EventCore.on('rup:paciente:fallecido', async (data) => {
-    try {
-        const lastSeguimiento = await SeguimientoPaciente.findOne({ 'paciente.id': data.prestacion.paciente.id }).sort({ createdAt: -1 });
-        if (lastSeguimiento) {
-            const prestacion = {
-                idPrestacion: data.prestacion._id,
-                tipoPrestacion: data.prestacion.solicitud.tipoPrestacion.term,
-                fecha: data.prestacion.createdAt
-            };
-            lastSeguimiento.ultimoEstado = { clave: 'fallecido', valor: prestacion.fecha };
-            lastSeguimiento.llamados.push(prestacion);
-            return await SeguimientoPacienteCtr.update(lastSeguimiento.id, lastSeguimiento, dataLog);
-        }
-    } catch (err) {
-        return err;
-    }
-});
